refactor(tenant-api): extract git command runner into gitops/gitcmd.go
writer.go (999 lines) was the largest hand-written file, mixing the low-level
git command runner with the high-level write orchestration. Move the runner —
3 timeout constants (defaultGit{Timeout,KillGrace,FetchTimeout}), 2 env-parsers
(gitTimeoutFromEnv, fetchTimeoutFromEnv), and gitCmd / gitCmdWithTimeout /
gitErr / clearStaleGitLocks / gitExec — into a new same-package file gitcmd.go.
writer.go drops to 842 lines; the #630/#638 timeout + lock-sweep machinery now
lives together.

Pure intra-package file move: same package gitops, no caller/test/signature
changes; dropped the now-unused io/fs + os/exec imports from writer.go.
Adversarial review confirms all 10 moved symbols are conserved (declared once),
the gitErr error-format strings are byte-identical, and go build/vet/test pass.

Self-Review-Pass-2: adversarial review: faithful move; 10 symbols conserved; gitops suite green
